Move patching tools to mocks.py.
Also improve documentation.
