+++ examples/users/users/settings_mocked.py
@@ -1,4 +1,4 @@
-from django_mock_queries.utils import monkey_patch_test_db
+from django_mock_queries.mocks import monkey_patch_test_db
 
 from users.settings import *
