+++ tests/__init__.py
@@ -1,3 +1,3 @@
-from django_mock_queries.utils import mock_django_setup
+from django_mock_queries.mocks import mock_django_setup
 
 mock_django_setup('tests.mock_settings')
